src/team-guidance/requirements.txt
langchain-core 
langchain-community 
langchain-openai
PyGithub
GitPython
loguru
pyyaml
PyGithub
pygit
tqdm
azure-search-documents
tqdm